requirements.txt
Django==1.5.1
# MySQL-python==1.2.4
South==0.7.6
Unidecode==0.04.12
amqp==1.0.11
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.28
celery==3.0.19
distribute==0.6.36
django-celery==3.0.17
django-haystack==2.1.0
django-social-auth==0.7.22
django-tastypie==0.9.14
django-uuslug==0.9.0
gunicorn==0.17.2
httplib2==0.8
kombu==2.5.10
mimeparse==0.1.3
oauth2==1.5.211
python-dateutil==2.1
python-openid==2.2.5
python-slugify==0.0.3
pytz==2013b
requests==1.2.0
six==1.3.0
wsgiref==0.1.2
